% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDhap.R
\name{LDhap}
\alias{LDhap}
\title{Query LDhap API}
\usage{
LDhap(snps, pop = "CEU", token = NULL, file = FALSE)
}
\arguments{
\item{snps}{list of between 1 - 30 variants, using an rsID or chromosome coordinate (e.g. "chr7:24966446")}

\item{pop}{a 1000 Genomes Project population, (e.g. YRI or CEU), multiple allowed, default = "CEU"}

\item{token}{LDlink provided user token, default = NULL, register for token at  \url{https://ldlink.nci.nih.gov/?tab=apiaccess}}

\item{file}{Optional character string naming a path and file for saving results.  If file = FALSE, no file will be generated, default = FALSE.}
}
\value{
a data frame
}
\description{
Query LDhap API
}
\examples{
\dontrun{LDhap(c("rs3", "rs4", "rs148890987"), "CEU", token = Sys.getenv("LDLINK_TOKEN"))}
\dontrun{LDhap("rs148890987", c("YRI", "CEU"), token = Sys.getenv("LDLINK_TOKEN"))}

}
