% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDmatrix.R
\name{LDmatrix}
\alias{LDmatrix}
\title{Query LDmatrix API}
\usage{
LDmatrix(snps, pop = "CEU", r2d = "r2", token = NULL, file = FALSE)
}
\arguments{
\item{snps}{list of between 2 - 1,000 variants, using an rsID or chromosome coordinate (e.g. "chr7:24966446")}

\item{pop}{pop, a particular population, (e.g. YRI or CEU), multiple allowed, default=CEU}

\item{r2d}{r2d, either "r2" for LD R-squared or "d" for LD D-prime, default="r2"}

\item{token}{token, LDlink provided user token, default = NULL, register for token at: https://ldlink.nci.nih.gov/?tab=apiaccess}

\item{file}{optional character string naming a path and file}
}
\value{
a data frame
}
\description{
Query LDmatrix API
}
\examples{
LDmatrix(c("rs3", "rs4", "rs148890987"), "YRI", "r2", "28da99809470")

}
