% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_chips.R
\name{list_chips}
\alias{list_chips}
\title{List available SNP Chip Arrays}
\usage{
list_chips()
}
\value{
a data frame listing the names and abbreviation codes for available SNP Chip Arrays from Illumina and Affymetrix
}
\description{
List available SNP Chip Arrays
}
\examples{
list_chips()
}
