% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_gtex_tissues.R
\name{list_gtex_tissues}
\alias{list_gtex_tissues}
\title{List GTEx tissues}
\usage{
list_gtex_tissues()
}
\value{
a data frame listing the GTEx tissues, their names and abbreviation codes
used as input for LDexpress.
}
\description{
List GTEx tissues
}
\examples{
list_gtex_tissues()
}
