% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_pop.R
\name{list_pop}
\alias{list_pop}
\title{Provides a data frame listing the available reference populations from
the 1000 Genomes Project.}
\usage{
list_pop()
}
\value{
a data frame listing the available reference populations, continental (ex: European, African, and Admixed American) and sub-populations (ex: Finnish, Gambian, and Peruvian)
}
\description{
Provides a data frame listing the available reference populations from
the 1000 Genomes Project.
}
\examples{
list_pop()

}
