% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPchip.R
\name{SNPchip}
\alias{SNPchip}
\title{Query SNPchip API}
\usage{
SNPchip(snps, chip = "ALL", token = NULL, file = FALSE)
}
\arguments{
\item{snps}{between 1 - 5,000 variants, using an rsID or chromosome coordinate (e.g. "chr7:24966446")}

\item{chip}{chip or arrays, platform code(s) for a SNP chip array, ALL_Illumina, ALL_Affy or ALL, default=ALL}

\item{token}{token, LDlink provided user token, default = NULL, register for token at: https://ldlink.nci.nih.gov/?tab=apiaccess}

\item{file}{optional character string naming a path and file}
}
\value{
a data frame
}
\description{
Query SNPchip API
}
\examples{
SNPchip(c("rs3", "rs4", "rs148890987"), "ALL", "28da99809470")
SNPchip(c("rs3", "rs4", "rs148890987"), c("A_CHB2", "A_SNP5.0"), "28da99809470")
SNPchip(c("rs3", "rs4", "rs148890987"), "ALL_Affy", "28da99809470")

}
