% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDlinkR-package.R
\docType{package}
\name{LDlinkR-package}
\alias{LDlinkR}
\alias{LDlinkR-package}
\title{LDlinkR: Calculating Linkage Disequilibrium (LD) in Human Population Groups of Interest}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides access to the LDlink API (<https://ldlink.nci.nih.gov/?tab=apiaccess>)
    using the R console.  This programmatic access facilitates researchers who are 
    interested in performing batch queries in 1000 Genomes Project data using LDlink. 
    LDlink is an interactive and powerful suite of web-based tools for querying germline 
    variants in human population groups of interest. For more details, please see 
    Machiela et al. (2015) <doi:10.1093/bioinformatics/btv402>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ldlink.nci.nih.gov}
  \item Report bugs at \url{https://github.com/CBIIT/LDlinkR/issues}
}

}
\author{
\strong{Maintainer}: Timothy A. Myers \email{myersta@mail.nih.gov}

Authors:
\itemize{
  \item Stephen J. Chanock \email{chanocks@mail.nih.gov}
  \item Mitchell J. Machiela \email{mitchell.machiela@nih.gov}
}

}
\keyword{internal}
