% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDlinkR-package.R
\docType{package}
\name{LDlinkR-package}
\alias{LDlinkR}
\alias{LDlinkR-package}
\title{LDlinkR: Access LDlink API with R}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Provides access to the LDlink API (https://ldlink.nci.nih.gov/?tab=apiaccess)
    using the R console.  This programmatic access facilitates researchers who are interested
    in performing batch queries in 1000 Genomes Project data using LDlink.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ldlink.nci.nih.gov}
}

}
\keyword{internal}
