% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDmatrix.R
\name{LDmatrix}
\alias{LDmatrix}
\title{Query LDmatrix API}
\usage{
LDmatrix(snps, pop = "CEU", r2d = "r2", token = NULL, file = FALSE)
}
\arguments{
\item{snps}{list of between 2 - 1,000 variants, using an rsID or chromosome coordinate (e.g. "chr7:24966446")}

\item{pop}{pop, a particular population, (e.g. YRI or CEU), multiple allowed, default = "CEU"}

\item{r2d}{r2d, either "r2" for LD R2 or "d" for LD D', default = "r2"}

\item{token}{LDlink provided user token, default = NULL, register for token at \url{https://ldlink.nci.nih.gov/?tab=apiaccess}}

\item{file}{Optional character string naming a path and file.  If file = FALSE, no file will be generated, default = FALSE.}
}
\value{
a data frame
}
\description{
Query LDmatrix API
}
\examples{
LDmatrix(c("rs3", "rs4", "rs148890987"), "YRI", "r2", token = Sys.getenv("LDLINK_TOKEN"))

}
