% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDpair.R
\name{LDpair}
\alias{LDpair}
\title{Query LDpair API}
\usage{
LDpair(var1, var2, pop = "CEU", token = NULL, output = "table",
  file = FALSE)
}
\arguments{
\item{var1}{the first RS number or genomic coordinate (e.g. "chr7:24966446")}

\item{var2}{the second RS number or genomic coordinate (e.g. "ch7:24966446")}

\item{pop}{a particular population(s), (e.g. YRI or CEU), multiple allowed, default = "CEU"}

\item{token}{LDlink provided user token, default = NULL, register for token at  \url{https://ldlink.nci.nih.gov/?tab=apiaccess}}

\item{output}{two output options available, "text" or "table", default = "table"}

\item{file}{Optional character string naming a path and file.  If file = FALSE, no file will be generated, default = FALSE.}
}
\value{
text or data frame, dependning on the output option
}
\description{
Query LDpair API
}
\examples{
LDpair(var1 = "rs3", var2 = "rs4", pop = "YRI", token = "faketoken123")
LDpair("rs3", "rs4", "YRI", "faketoken123", "text")

}
