% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDpop.R
\name{LDpop}
\alias{LDpop}
\title{Query LDpop API}
\usage{
LDpop(var1, var2, pop = "CEU", r2d = "r2", token = NULL,
  file = FALSE)
}
\arguments{
\item{var1}{the first RS number or genomic coordinate (e.g. "chr7:24966446")}

\item{var2}{the second RS number or genomic coordinate (e.g. "ch7:24966446")}

\item{pop}{a particular population(s), (e.g. YRI or CEU), multiple allowed, default=CEU}

\item{r2d}{either "r2" for LD R-squared or "d" for LD D-prime, default="r2"}

\item{token}{LDlink provided user token, default = NULL, register for token at: https://ldlink.nci.nih.gov/?tab=apiaccess}

\item{file}{optional character string naming a path and file}
}
\value{
a data frame
}
\description{
Query LDpop API
}
\examples{
LDpop(var1 = "rs3", var2 = "rs4", pop = "YRI", r2d = "r2", token = "28da99809470")

}
