% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDproxy.R
\name{LDproxy}
\alias{LDproxy}
\title{Query LDproxy API}
\usage{
LDproxy(snp, pop = "CEU", r2d = "r2", token = NULL, file = FALSE)
}
\arguments{
\item{snp}{an rsID or chromosome coordinate (e.g. "chr7:24966446"), one per query}

\item{pop}{a particular population, (e.g. YRI or CEU), multiple allowed, default=CEU}

\item{r2d}{r2d, either "r2" for LD R-squared or "d" for LD D-prime, default="r2"}

\item{token}{token, LDlink provided user token, default = NULL, register for token at: https://ldlink.nci.nih.gov/?tab=apiaccess}

\item{file}{optional character string naming a path and file}
}
\value{
a data frame
}
\description{
Query LDproxy API
}
\examples{
df_proxy <- LDproxy("rs456", "YRI", "r2", "faketoken123")

}
