% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_gtex_tissues.R
\name{list_gtex_tissues}
\alias{list_gtex_tissues}
\title{Provides a data frame listing the GTEx full names, `LDexpress` full names
(without spaces) and acceptable abbreviation codes of the 54 non-diseased
tissue sites collected for the GTEx Portal and used as input for the
`LDexpress` function.}
\usage{
list_gtex_tissues()
}
\value{
a data frame listing the GTEx tissues, their names and abbreviation codes
used as input for LDexpress.
}
\description{
Provides a data frame listing the GTEx full names, `LDexpress` full names
(without spaces) and acceptable abbreviation codes of the 54 non-diseased
tissue sites collected for the GTEx Portal and used as input for the
`LDexpress` function.
}
\examples{
list_gtex_tissues()
}
